Create authorization and reflect token expiration in UI

+++ backend/routes/posts.js
@@ -39,7 +39,8 @@ router.post("", checkAuth, multer({storage: storage}).single("image") ,(req, res
   const post = new Post({
     title: req.body.title,
     content: req.body.content,
-    imagePath: url + "/images/" + req.file.filename
+    imagePath: url + "/images/" + req.file.filename,
+    creator: req.userData.userId
   });
 
   //save is a mongooose method
@@ -117,22 +118,30 @@ router.put("/:id", checkAuth, multer({storage: storage}).single("image"), (req, 
     _id: req.body.id,
     title: req.body.title,
     content: req.body.content,
-    imagePath: imagePath
+    imagePath: imagePath,
+    creator: req.userData.userId
   });
 
-  Post.updateOne({_id: req.params.id}, post).then(result => {
-    res.status(200).json({ message: 'Update successful'})
+  Post.updateOne({_id: req.params.id, creator: req.userData.userId}, post).then(result => {
+    if(result.nModified > 0){
+      res.status(200).json({ message: 'Update successful'})
+    }else{
+      res.status(401).json({ message: 'Not Authorized'})
+    }
+
   })
 })
 
 //the colon shows dynamic path
 router.delete("/:id", checkAuth, (req, res, next) => {
 
-  Post.deleteOne({ _id: req.params.id}).then(result =>{
-    console.log(result);
-    res.status(200).json({
-      message: 'document deleted'
-    });
+  Post.deleteOne({ _id: req.params.id, creator: req.userData.userId}).then(result =>{
+
+    if(result.n > 0){
+      res.status(200).json({ message: 'Deletion successful'})
+    }else{
+      res.status(401).json({ message: 'Not Authorized'})
+    }
   });
 
 });
